Add tests, examples, pyproject.toml, and fix bugs (phase 2)
Tests (94 passing):
- tests/test_vision.py: rubric categories, scoring bands, prompt evaluator
- tests/test_geometry.py: PatternPiece CRUD, encoder, scaler, grade rules
- tests/test_output.py: SVG/PDF rendering with feature toggles
- tests/test_pipeline.py: end-to-end vision→encode→scale→export wiring

Examples:
- examples/quick_start.py: minimal pipeline in ~20 lines
- examples/pants_pullon.py: multi-piece pants with full export

Package config:
- pyproject.toml with dependencies and optional extras

Bug fixes found during testing:
- scaler.py: np.array dtype int64→float64 for displacement arithmetic
- svg_writer.py: polygon points as list of tuples, not string (svgwrite API)

+++ pattern_geometry/scaler.py
@@ -406,7 +406,7 @@ class PatternScaler:
         if not points:
             return points, []
 
-        points_arr = np.array(points)
+        points_arr = np.array(points, dtype=np.float64)
         movements = []
 
         # Identify boundary regions by their relative position
